fix: Duplicate layout use
