+++ frontend/src/routeTree.gen.ts
@@ -22,15 +22,14 @@ import { Route as AuthenticatedAgentsIndexRouteImport } from './routes/_authenti
 import { Route as AuthenticatedAdminIndexRouteImport } from './routes/_authenticated/admin/index'
 import { Route as AuthenticatedSettingsSecurityRouteImport } from './routes/_authenticated/settings/security'
 import { Route as AuthenticatedSettingsProfileRouteImport } from './routes/_authenticated/settings/profile'
+import { Route as AuthenticatedApplicationsIdRouteImport } from './routes/_authenticated/applications/$id'
 import { Route as AuthenticatedAdminUsersRouteImport } from './routes/_authenticated/admin/users'
 import { Route as AuthenticatedAdminSystemInfoRouteImport } from './routes/_authenticated/admin/system-info'
 import { Route as AuthenticatedAdminOidcProvidersRouteImport } from './routes/_authenticated/admin/oidc-providers'
 import { Route as AuthenticatedAdminAuditLogRouteImport } from './routes/_authenticated/admin/audit-log'
 import { Route as AuthenticatedApplicationsIdIndexRouteImport } from './routes/_authenticated/applications/$id/index'
-import { Route as AuthenticatedApplicationsIdSettingsRouteImport } from './routes/_authenticated/applications/$id/settings'
 import { Route as AuthenticatedApplicationsIdDetailsRouteImport } from './routes/_authenticated/applications/$id/details'
 import { Route as AuthenticatedApplicationsIdSettingsIndexRouteImport } from './routes/_authenticated/applications/$id/settings/index'
-import { Route as AuthenticatedApplicationsIdDetailsIndexRouteImport } from './routes/_authenticated/applications/$id/details/index'
 import { Route as AuthenticatedApplicationsIdSettingsImagePollingRouteImport } from './routes/_authenticated/applications/$id/settings/image-polling'
 import { Route as AuthenticatedApplicationsIdSettingsGeneralRouteImport } from './routes/_authenticated/applications/$id/settings/general'
 
@@ -105,6 +104,12 @@ const AuthenticatedSettingsProfileRoute =
     path: '/profile',
     getParentRoute: () => AuthenticatedSettingsRoute,
   } as any)
+const AuthenticatedApplicationsIdRoute =
+  AuthenticatedApplicationsIdRouteImport.update({
+    id: '/applications/$id',
+    path: '/applications/$id',
+    getParentRoute: () => AuthenticatedRoute,
+  } as any)
 const AuthenticatedAdminUsersRoute = AuthenticatedAdminUsersRouteImport.update({
   id: '/users',
   path: '/users',
@@ -130,45 +135,33 @@ const AuthenticatedAdminAuditLogRoute =
   } as any)
 const AuthenticatedApplicationsIdIndexRoute =
   AuthenticatedApplicationsIdIndexRouteImport.update({
-    id: '/applications/$id/',
-    path: '/applications/$id/',
-    getParentRoute: () => AuthenticatedRoute,
-  } as any)
-const AuthenticatedApplicationsIdSettingsRoute =
-  AuthenticatedApplicationsIdSettingsRouteImport.update({
-    id: '/applications/$id/settings',
-    path: '/applications/$id/settings',
-    getParentRoute: () => AuthenticatedRoute,
+    id: '/',
+    path: '/',
+    getParentRoute: () => AuthenticatedApplicationsIdRoute,
   } as any)
 const AuthenticatedApplicationsIdDetailsRoute =
   AuthenticatedApplicationsIdDetailsRouteImport.update({
-    id: '/applications/$id/details',
-    path: '/applications/$id/details',
-    getParentRoute: () => AuthenticatedRoute,
+    id: '/details',
+    path: '/details',
+    getParentRoute: () => AuthenticatedApplicationsIdRoute,
   } as any)
 const AuthenticatedApplicationsIdSettingsIndexRoute =
   AuthenticatedApplicationsIdSettingsIndexRouteImport.update({
-    id: '/',
-    path: '/',
-    getParentRoute: () => AuthenticatedApplicationsIdSettingsRoute,
-  } as any)
-const AuthenticatedApplicationsIdDetailsIndexRoute =
-  AuthenticatedApplicationsIdDetailsIndexRouteImport.update({
-    id: '/',
-    path: '/',
-    getParentRoute: () => AuthenticatedApplicationsIdDetailsRoute,
+    id: '/settings/',
+    path: '/settings/',
+    getParentRoute: () => AuthenticatedApplicationsIdRoute,
   } as any)
 const AuthenticatedApplicationsIdSettingsImagePollingRoute =
   AuthenticatedApplicationsIdSettingsImagePollingRouteImport.update({
-    id: '/image-polling',
-    path: '/image-polling',
-    getParentRoute: () => AuthenticatedApplicationsIdSettingsRoute,
+    id: '/settings/image-polling',
+    path: '/settings/image-polling',
+    getParentRoute: () => AuthenticatedApplicationsIdRoute,
   } as any)
 const AuthenticatedApplicationsIdSettingsGeneralRoute =
   AuthenticatedApplicationsIdSettingsGeneralRouteImport.update({
-    id: '/general',
-    path: '/general',
-    getParentRoute: () => AuthenticatedApplicationsIdSettingsRoute,
+    id: '/settings/general',
+    path: '/settings/general',
+    getParentRoute: () => AuthenticatedApplicationsIdRoute,
   } as any)
 
 export interface FileRoutesByFullPath {
@@ -180,6 +173,7 @@ export interface FileRoutesByFullPath {
   '/admin/oidc-providers': typeof AuthenticatedAdminOidcProvidersRoute
   '/admin/system-info': typeof AuthenticatedAdminSystemInfoRoute
   '/admin/users': typeof AuthenticatedAdminUsersRoute
+  '/applications/$id': typeof AuthenticatedApplicationsIdRouteWithChildren
   '/settings/profile': typeof AuthenticatedSettingsProfileRoute
   '/settings/security': typeof AuthenticatedSettingsSecurityRoute
   '/admin/': typeof AuthenticatedAdminIndexRoute
@@ -188,12 +182,10 @@ export interface FileRoutesByFullPath {
   '/notifications/': typeof AuthenticatedNotificationsIndexRoute
   '/repositories/': typeof AuthenticatedRepositoriesIndexRoute
   '/settings/': typeof AuthenticatedSettingsIndexRoute
-  '/applications/$id/details': typeof AuthenticatedApplicationsIdDetailsRouteWithChildren
-  '/applications/$id/settings': typeof AuthenticatedApplicationsIdSettingsRouteWithChildren
+  '/applications/$id/details': typeof AuthenticatedApplicationsIdDetailsRoute
   '/applications/$id/': typeof AuthenticatedApplicationsIdIndexRoute
   '/applications/$id/settings/general': typeof AuthenticatedApplicationsIdSettingsGeneralRoute
   '/applications/$id/settings/image-polling': typeof AuthenticatedApplicationsIdSettingsImagePollingRoute
-  '/applications/$id/details/': typeof AuthenticatedApplicationsIdDetailsIndexRoute
   '/applications/$id/settings/': typeof AuthenticatedApplicationsIdSettingsIndexRoute
 }
 export interface FileRoutesByTo {
@@ -211,10 +203,10 @@ export interface FileRoutesByTo {
   '/notifications': typeof AuthenticatedNotificationsIndexRoute
   '/repositories': typeof AuthenticatedRepositoriesIndexRoute
   '/settings': typeof AuthenticatedSettingsIndexRoute
+  '/applications/$id/details': typeof AuthenticatedApplicationsIdDetailsRoute
   '/applications/$id': typeof AuthenticatedApplicationsIdIndexRoute
   '/applications/$id/settings/general': typeof AuthenticatedApplicationsIdSettingsGeneralRoute
   '/applications/$id/settings/image-polling': typeof AuthenticatedApplicationsIdSettingsImagePollingRoute
-  '/applications/$id/details': typeof AuthenticatedApplicationsIdDetailsIndexRoute
   '/applications/$id/settings': typeof AuthenticatedApplicationsIdSettingsIndexRoute
 }
 export interface FileRoutesById {
@@ -228,6 +220,7 @@ export interface FileRoutesById {
   '/_authenticated/admin/oidc-providers': typeof AuthenticatedAdminOidcProvidersRoute
   '/_authenticated/admin/system-info': typeof AuthenticatedAdminSystemInfoRoute
   '/_authenticated/admin/users': typeof AuthenticatedAdminUsersRoute
+  '/_authenticated/applications/$id': typeof AuthenticatedApplicationsIdRouteWithChildren
   '/_authenticated/settings/profile': typeof AuthenticatedSettingsProfileRoute
   '/_authenticated/settings/security': typeof AuthenticatedSettingsSecurityRoute
   '/_authenticated/admin/': typeof AuthenticatedAdminIndexRoute
@@ -236,12 +229,10 @@ export interface FileRoutesById {
   '/_authenticated/notifications/': typeof AuthenticatedNotificationsIndexRoute
   '/_authenticated/repositories/': typeof AuthenticatedRepositoriesIndexRoute
   '/_authenticated/settings/': typeof AuthenticatedSettingsIndexRoute
-  '/_authenticated/applications/$id/details': typeof AuthenticatedApplicationsIdDetailsRouteWithChildren
-  '/_authenticated/applications/$id/settings': typeof AuthenticatedApplicationsIdSettingsRouteWithChildren
+  '/_authenticated/applications/$id/details': typeof AuthenticatedApplicationsIdDetailsRoute
   '/_authenticated/applications/$id/': typeof AuthenticatedApplicationsIdIndexRoute
   '/_authenticated/applications/$id/settings/general': typeof AuthenticatedApplicationsIdSettingsGeneralRoute
   '/_authenticated/applications/$id/settings/image-polling': typeof AuthenticatedApplicationsIdSettingsImagePollingRoute
-  '/_authenticated/applications/$id/details/': typeof AuthenticatedApplicationsIdDetailsIndexRoute
   '/_authenticated/applications/$id/settings/': typeof AuthenticatedApplicationsIdSettingsIndexRoute
 }
 export interface FileRouteTypes {
@@ -255,6 +246,7 @@ export interface FileRouteTypes {
     | '/admin/oidc-providers'
     | '/admin/system-info'
     | '/admin/users'
+    | '/applications/$id'
     | '/settings/profile'
     | '/settings/security'
     | '/admin/'
@@ -264,11 +256,9 @@ export interface FileRouteTypes {
     | '/repositories/'
     | '/settings/'
     | '/applications/$id/details'
-    | '/applications/$id/settings'
     | '/applications/$id/'
     | '/applications/$id/settings/general'
     | '/applications/$id/settings/image-polling'
-    | '/applications/$id/details/'
     | '/applications/$id/settings/'
   fileRoutesByTo: FileRoutesByTo
   to:
@@ -286,10 +276,10 @@ export interface FileRouteTypes {
     | '/notifications'
     | '/repositories'
     | '/settings'
+    | '/applications/$id/details'
     | '/applications/$id'
     | '/applications/$id/settings/general'
     | '/applications/$id/settings/image-polling'
-    | '/applications/$id/details'
     | '/applications/$id/settings'
   id:
     | '__root__'
@@ -302,6 +292,7 @@ export interface FileRouteTypes {
     | '/_authenticated/admin/oidc-providers'
     | '/_authenticated/admin/system-info'
     | '/_authenticated/admin/users'
+    | '/_authenticated/applications/$id'
     | '/_authenticated/settings/profile'
     | '/_authenticated/settings/security'
     | '/_authenticated/admin/'
@@ -311,11 +302,9 @@ export interface FileRouteTypes {
     | '/_authenticated/repositories/'
     | '/_authenticated/settings/'
     | '/_authenticated/applications/$id/details'
-    | '/_authenticated/applications/$id/settings'
     | '/_authenticated/applications/$id/'
     | '/_authenticated/applications/$id/settings/general'
     | '/_authenticated/applications/$id/settings/image-polling'
-    | '/_authenticated/applications/$id/details/'
     | '/_authenticated/applications/$id/settings/'
   fileRoutesById: FileRoutesById
 }
@@ -417,6 +406,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AuthenticatedSettingsProfileRouteImport
       parentRoute: typeof AuthenticatedSettingsRoute
     }
+    '/_authenticated/applications/$id': {
+      id: '/_authenticated/applications/$id'
+      path: '/applications/$id'
+      fullPath: '/applications/$id'
+      preLoaderRoute: typeof AuthenticatedApplicationsIdRouteImport
+      parentRoute: typeof AuthenticatedRoute
+    }
     '/_authenticated/admin/users': {
       id: '/_authenticated/admin/users'
       path: '/users'
@@ -447,52 +443,38 @@ declare module '@tanstack/react-router' {
     }
     '/_authenticated/applications/$id/': {
       id: '/_authenticated/applications/$id/'
-      path: '/applications/$id'
+      path: '/'
       fullPath: '/applications/$id/'
       preLoaderRoute: typeof AuthenticatedApplicationsIdIndexRouteImport
-      parentRoute: typeof AuthenticatedRoute
-    }
-    '/_authenticated/applications/$id/settings': {
-      id: '/_authenticated/applications/$id/settings'
-      path: '/applications/$id/settings'
-      fullPath: '/applications/$id/settings'
-      preLoaderRoute: typeof AuthenticatedApplicationsIdSettingsRouteImport
-      parentRoute: typeof AuthenticatedRoute
+      parentRoute: typeof AuthenticatedApplicationsIdRoute
     }
     '/_authenticated/applications/$id/details': {
       id: '/_authenticated/applications/$id/details'
-      path: '/applications/$id/details'
+      path: '/details'
       fullPath: '/applications/$id/details'
       preLoaderRoute: typeof AuthenticatedApplicationsIdDetailsRouteImport
-      parentRoute: typeof AuthenticatedRoute
+      parentRoute: typeof AuthenticatedApplicationsIdRoute
     }
     '/_authenticated/applications/$id/settings/': {
       id: '/_authenticated/applications/$id/settings/'
-      path: '/'
+      path: '/settings'
       fullPath: '/applications/$id/settings/'
       preLoaderRoute: typeof AuthenticatedApplicationsIdSettingsIndexRouteImport
-      parentRoute: typeof AuthenticatedApplicationsIdSettingsRoute
-    }
-    '/_authenticated/applications/$id/details/': {
-      id: '/_authenticated/applications/$id/details/'
-      path: '/'
-      fullPath: '/applications/$id/details/'
-      preLoaderRoute: typeof AuthenticatedApplicationsIdDetailsIndexRouteImport
-      parentRoute: typeof AuthenticatedApplicationsIdDetailsRoute
+      parentRoute: typeof AuthenticatedApplicationsIdRoute
     }
     '/_authenticated/applications/$id/settings/image-polling': {
       id: '/_authenticated/applications/$id/settings/image-polling'
-      path: '/image-polling'
+      path: '/settings/image-polling'
       fullPath: '/applications/$id/settings/image-polling'
       preLoaderRoute: typeof AuthenticatedApplicationsIdSettingsImagePollingRouteImport
-      parentRoute: typeof AuthenticatedApplicationsIdSettingsRoute
+      parentRoute: typeof AuthenticatedApplicationsIdRoute
     }
     '/_authenticated/applications/$id/settings/general': {
       id: '/_authenticated/applications/$id/settings/general'
-      path: '/general'
+      path: '/settings/general'
       fullPath: '/applications/$id/settings/general'
       preLoaderRoute: typeof AuthenticatedApplicationsIdSettingsGeneralRouteImport
-      parentRoute: typeof AuthenticatedApplicationsIdSettingsRoute
+      parentRoute: typeof AuthenticatedApplicationsIdRoute
     }
   }
 }
@@ -533,29 +515,20 @@ const AuthenticatedSettingsRouteWithChildren =
     AuthenticatedSettingsRouteChildren,
   )
 
-interface AuthenticatedApplicationsIdDetailsRouteChildren {
-  AuthenticatedApplicationsIdDetailsIndexRoute: typeof AuthenticatedApplicationsIdDetailsIndexRoute
-}
-
-const AuthenticatedApplicationsIdDetailsRouteChildren: AuthenticatedApplicationsIdDetailsRouteChildren =
-  {
-    AuthenticatedApplicationsIdDetailsIndexRoute:
-      AuthenticatedApplicationsIdDetailsIndexRoute,
-  }
-
-const AuthenticatedApplicationsIdDetailsRouteWithChildren =
-  AuthenticatedApplicationsIdDetailsRoute._addFileChildren(
-    AuthenticatedApplicationsIdDetailsRouteChildren,
-  )
-
-interface AuthenticatedApplicationsIdSettingsRouteChildren {
+interface AuthenticatedApplicationsIdRouteChildren {
+  AuthenticatedApplicationsIdDetailsRoute: typeof AuthenticatedApplicationsIdDetailsRoute
+  AuthenticatedApplicationsIdIndexRoute: typeof AuthenticatedApplicationsIdIndexRoute
   AuthenticatedApplicationsIdSettingsGeneralRoute: typeof AuthenticatedApplicationsIdSettingsGeneralRoute
   AuthenticatedApplicationsIdSettingsImagePollingRoute: typeof AuthenticatedApplicationsIdSettingsImagePollingRoute
   AuthenticatedApplicationsIdSettingsIndexRoute: typeof AuthenticatedApplicationsIdSettingsIndexRoute
 }
 
-const AuthenticatedApplicationsIdSettingsRouteChildren: AuthenticatedApplicationsIdSettingsRouteChildren =
+const AuthenticatedApplicationsIdRouteChildren: AuthenticatedApplicationsIdRouteChildren =
   {
+    AuthenticatedApplicationsIdDetailsRoute:
+      AuthenticatedApplicationsIdDetailsRoute,
+    AuthenticatedApplicationsIdIndexRoute:
+      AuthenticatedApplicationsIdIndexRoute,
     AuthenticatedApplicationsIdSettingsGeneralRoute:
       AuthenticatedApplicationsIdSettingsGeneralRoute,
     AuthenticatedApplicationsIdSettingsImagePollingRoute:
@@ -564,37 +537,32 @@ const AuthenticatedApplicationsIdSettingsRouteChildren: AuthenticatedApplication
       AuthenticatedApplicationsIdSettingsIndexRoute,
   }
 
-const AuthenticatedApplicationsIdSettingsRouteWithChildren =
-  AuthenticatedApplicationsIdSettingsRoute._addFileChildren(
-    AuthenticatedApplicationsIdSettingsRouteChildren,
+const AuthenticatedApplicationsIdRouteWithChildren =
+  AuthenticatedApplicationsIdRoute._addFileChildren(
+    AuthenticatedApplicationsIdRouteChildren,
   )
 
 interface AuthenticatedRouteChildren {
   AuthenticatedAdminRoute: typeof AuthenticatedAdminRouteWithChildren
   AuthenticatedSettingsRoute: typeof AuthenticatedSettingsRouteWithChildren
   AuthenticatedIndexRoute: typeof AuthenticatedIndexRoute
+  AuthenticatedApplicationsIdRoute: typeof AuthenticatedApplicationsIdRouteWithChildren
   AuthenticatedAgentsIndexRoute: typeof AuthenticatedAgentsIndexRoute
   AuthenticatedApplicationsIndexRoute: typeof AuthenticatedApplicationsIndexRoute
   AuthenticatedNotificationsIndexRoute: typeof AuthenticatedNotificationsIndexRoute
   AuthenticatedRepositoriesIndexRoute: typeof AuthenticatedRepositoriesIndexRoute
-  AuthenticatedApplicationsIdDetailsRoute: typeof AuthenticatedApplicationsIdDetailsRouteWithChildren
-  AuthenticatedApplicationsIdSettingsRoute: typeof AuthenticatedApplicationsIdSettingsRouteWithChildren
-  AuthenticatedApplicationsIdIndexRoute: typeof AuthenticatedApplicationsIdIndexRoute
 }
 
 const AuthenticatedRouteChildren: AuthenticatedRouteChildren = {
   AuthenticatedAdminRoute: AuthenticatedAdminRouteWithChildren,
   AuthenticatedSettingsRoute: AuthenticatedSettingsRouteWithChildren,
   AuthenticatedIndexRoute: AuthenticatedIndexRoute,
+  AuthenticatedApplicationsIdRoute:
+    AuthenticatedApplicationsIdRouteWithChildren,
   AuthenticatedAgentsIndexRoute: AuthenticatedAgentsIndexRoute,
   AuthenticatedApplicationsIndexRoute: AuthenticatedApplicationsIndexRoute,
   AuthenticatedNotificationsIndexRoute: AuthenticatedNotificationsIndexRoute,
   AuthenticatedRepositoriesIndexRoute: AuthenticatedRepositoriesIndexRoute,
-  AuthenticatedApplicationsIdDetailsRoute:
-    AuthenticatedApplicationsIdDetailsRouteWithChildren,
-  AuthenticatedApplicationsIdSettingsRoute:
-    AuthenticatedApplicationsIdSettingsRouteWithChildren,
-  AuthenticatedApplicationsIdIndexRoute: AuthenticatedApplicationsIdIndexRoute,
 }
 
 const AuthenticatedRouteWithChildren = AuthenticatedRoute._addFileChildren(
